Update patch set 2
Patch Set 2: Code-Review-1

This looks great

Two minor things

elements without anything to pull down are showing a ">" that doesn't expand anything; I think these need to not to set a compoundExpand?

I think the label that switches is supposed to be to the right of a switch, with static text to it's left; so it's like 

 <what switch does> <o> <on|off>

I don't think this even needs an on/off label; so it can just say "Show retried jobs" and if the slider has no label I think that should show a tick when enabled.  To me this clearly communicates what it is, and stops the label jumping around when flipped

Patch-set: 2
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1