Fix parsing of zone from path in executor API
The get() method of the executor API receives the full path to a build
request. In case of a zoned request the zone wasn't correctly derived
from the given path so that the resulting "zone" still contained the
request's UUID.

This lead to the creation of an `ExecutorQueue` for each build request
through the default factory of the `DefaultKeyDict`.

Change-Id: I9aafcee0ea24b45d095e35e242c24639dc149536