Don't use the AnsibleJob in the nodepool client
This change follows up on a few TODOs left by the lock/unlock nodes on
executor change.

When locking the nodes on the executor we used the AnsibleJob as a
replacement for the old build parameter that was provided to the
nodepool client methods as they were originally called by the scheduler.

However, the AnsibleJob class should only be used internally by the
executor, so we now provide all parameters directly to the nodepool
methods.

This also annotates the logger in the updated nodepool client methods
and fixes an outdated method signature in
test_scheduler.TestSemaphore.test_semaphore_zk_error.

Change-Id: I2d1005f69904c6ace8f79523133f382af0024c52