Update patch set 11
Patch Set 11:

(1 comment)

> Patch Set 9:
> 
> I see, perhaps we don't need to change the API, the client should already know what is the offset, and the size can be calculated using Array.length.

The problem I see is that with the current API, the "limit" argument cannot be omitted. We could set limit to a default arbitrarily high value in the GUI, or change the web API to remove the limit in the query if set to -1 for example, and pull all the results into the React App, then do the pagination in the javascript. I am afraid of the performance hit on zuul-web though, if it has to fetch all the builds from the db then send them to the client.

This is why I favored handling the pagination in zuul-web.

> 
> Otherwise we could split the api change so that the ui can be tested.

I can create a patch for the API change, but that would break the GUI if the GUI change isn't merged right after it, wouldn't it?

Patch-set: 11