Query gerrit files endpoint
The /changes/x/revisions/y/files endpoint in the Gerrit REST API
can be used to obtain a list of files modified not only by the
commit itself but also files which have changed compared to the
target branch if the commit is a merge commit.  No other methods
of file listing (SSH or the plain /changes endpoint) provide this
information.  This means that changes to zuul.yaml files that
arrive via merge commits which themselves don't alter the file
are invisible to Zuul.

The Gerrit project's workflow of merging up changes from stable
branches to master is particularly suceptible to this problem.

To correct this, query the files endpoint explicitly.  The test
framework is updated to support creating merge commits and
simulating the Gerrit query results as well.  Note that these
methods are not completely fleshed out -- they are only sufficient
to support this simple case.

Change-Id: I6eb91abd46f0d68ef1e5c2f500ff16e8d094693e