Don't use executor.builds when processing build result events
The executor client still holds a list of local builds objects which is
used in various places. One use case is to look up necessary
information of the original build when a build result event is handled.

Using such a local list won't work with multiple schedulers in place. As
a first step we will avoid using this list for handling build result
events and instead provide all necessary information to the build result
itself and look up the remaining information from the pipeline directly.

Change-Id: I9c4e573de2ce63259ec6cfb7d69c2f5be48f33ef