Update patch set 10
Patch Set 10: Code-Review-1

The (X) in the entry field also clears the filters:

* Go to builds page for the zuul tenant
* Job is selected by default; type in "zuul-" click "zuul-tox-py38"
* Page refreshes to show just those jobs
* Type in "zuul-foo"
* Change your mind and click (X) to the right of "zuul-foo"
* Existing search for zuul-tox-py38 is cleared and page resets to default

Especially since there's an (x) on the card, I think the user expectation is that the (X) in the search entry should only apply to the entry field.

It's possible this bug is introduced in a parent change, I just didn't think to test that with those fields, but it seems much more likely for job input.

Patch-set: 10
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1