Update patch set 10
Patch Set 10:

> Patch Set 10: Code-Review-1
> 
> The (X) in the entry field also clears the filters:
> 
> * Go to builds page for the zuul tenant
> * Job is selected by default; type in "zuul-" click "zuul-tox-py38"
> * Page refreshes to show just those jobs
> * Type in "zuul-foo"
> * Change your mind and click (X) to the right of "zuul-foo"
> * Existing search for zuul-tox-py38 is cleared and page resets to default
> 
> Especially since there's an (x) on the card, I think the user expectation is that the (X) in the search entry should only apply to the entry field.
> 
> It's possible this bug is introduced in a parent change, I just didn't think to test that with those fields, but it seems much more likely for job input.

This seems to be the intended behavior in Patternfly: you can reproduce what you describe in the documentation's example at https://www.patternfly.org/v4/components/select#multiple

There's an onClear property in the select, it should be possible to modify that function to check first if there's text in the select field and if so, remove it; otherwise clear all the selections for the specific category. I'm not sure it's a good UX practice though to have different behaviors depending on what's in the component.

As an aside, my first reflex would have been to hit backspace to remove "zuul-foo" from the search field, but I can see how that button can be misleading!

Patch-set: 10