Update patch set 10
Patch Set 10:

If that's intended, that's... unfortunate.  I feel it's very misleading.  Perhaps part of the issue is the way we have constructed the page to have chips below the entry field.

TBH, I'd prefer to fix this by keeping the chips where they are and altering the big (X) to only clear partial entries.  But if that doesn't work/isn't acceptable, then we may need to restructure our page to be more like the example and put the chips in the input field so it's more clear that the "clear the input field" button might clear the chips.

Patch-set: 10
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>