Remove unneeded scheduler.zk_nodepool object
The scheduler has a Nodeool object, and the Nodepool object has
a ZooKeeperNodepool object.  Separately, the scheduler also has a
standalone ZooKeeperNodepool object.  Rather than having a second
zk_nodepool object, just reach into Nodepool object and use its
zk_nodepool object directly.

This is more important now that ZooKeeperNodepool maintains a
node request cache (and will also maintain a node cache in a future
change).  This means that the scheduler was keeping two in-memory
caches, which is extra work being performed.

Because one of the zk_nodepool objects was being used to generate
nodes provisioned events, and the other was being used to process
them, if their caches weren't in sync, the scheduler could end up
marking node requests as failed when they actually succeeded.

The dual cache issue is why we saw this issue in tests, but the
same issue would be present with multiple schedulers too, so we
also update the getNodeRequest method to make the cache optional.
We bypass the cache where we must be certain we have the most
up to date info.

Change-Id: I89242a01f656abce143bfb991670d452deae8b72