Fix test race with node allocation
In order to determine if the system is settled, the test framework
checks whether all node requests have been fulfilled (and, implicitly,
that Zuul has seen and processed that fulfillment).  That sequence
is now:

1. Request state in ZK is set to fulfilled by nodepool
2. Scheduler receives watch that request is fulfilled
3. Internal cache is updated to reflect new state
4. NodesProvisionedEvent is added to ZK result event queue

There is a window between 3 and 4 where, to an external observer, the
system looks the same as after 4.  We used to have an internal list of
pending requests in the scheduler which we would not clear until after
4, however that has been removed since it doesn't make sense with
multiple schedulers.

To resolve this, we wrap steps 3 and 4 in a lock so they act as a
critical section, and then in the test framework, we grab that same lock
and check both the internal cache (3) and the result event queue(4) to
determine if we're between 3 and 4 (not settled) or after 4 (settled).

Change-Id: Ib6d0ad826cc4d3fad9d1b59434971f48bab7d23a