Handle de-serializing NodeRequests without requestor_data
Our internal cache is updated for every node request in zookeeper,
but some of those requests aren't from us and don't have requestor_data
in the format we expect.  We're not going to do anything with those
instances, but before we get to that point, we do need to be able
to deserialize them so we can determine they don't belong to us.

Handle the cases where requestor_data is None or is a dictionary
without all the keys we expect.

Nodepool min-ready requests can cause this.

Change-Id: If6b6ad156c4a1306145056add3db1df3467beacd