Don't use executor.builds to find out if tests are settled
This changes the __areAllBuildsWaiting() and
__haveAllBuildsReported() test methods to not rely on the local
executor.builds list but rather use the build requests stored in
ZooKeeper to find out if the system is settled/stable.

This also removes an unnecessary waitUntilSettled() call from
test_data_return_child_from_retried_paused_job which seems to be doing
too much now (most probably due to some timing issue).

I've tested this fix on the master state and there this
waitUntilSettled() call is also not needed.

Change-Id: I050ac965142e3ba87bd180257493aea0882056b6