Move time database to SQL
To support multiple schedulers, move the time database into the
SQL database.

We could use the existing build table, however performance can
be somewhat unpredictable (at least with postgres, we observed
a range of 2ms to 400ms to perform the necessary query depending
on the project/job selected and the contents of the database).
So instead, we create a new table dedicated to this task, with
one row per tenant-project-branch-job combination.  This row
has 10 slots for times and rotates the values through accordingly.

There is a small chance we may collide when updating this table;
that's not important so we don't attempt to recover (we will just
lose that time update).

Nothing in the scheduler uses the state directory anymore, so clean
up the docs around that.  The executor still has a state dir where
it may install ansible-related files.

Change-Id: Id3c64be7a05c9edc849e698200411ad436a1334d