Fix race in test_data_return_child_from_retried_paused_job
There seems to be a race where the paused job (paused-data-return-vars)
is sometimes kept on hold and thus doesn't continue during the
waitUntilSettled() because the executor_server.hold_jobs_in_build flag
is still set to True. In this case the following release of the
print-data-return-vars job doesn't work because this build is not yet
executed because it's waiting on the paused-data-return-vars job.

Disabling the hold_jobs_in_build before the waitUntilSettled() call
seems to do the trick and allows the job to continue as expected.

Change-Id: I1b026434f24df62b94ea66efed584b8ab226ca7e