Fix waitUntilSettled nodepool race
Change cebebf0f4249a1b0cddc2f501cd245c5c9766dc2 closed most of a
race condition regarding determining when Zuul was settled in unit
tests.  It relies on verifying that the scheduler's cache of nodepool
requests matches ZK, and in doing so, it uses a method that retrieves
the cached request object.  But that method will perform a ZK lookup
if it encounters a cache miss.  That would make it appear the cache
is in sync when it is not.

To correct this, inspect the cache dictionary directly within the
unit tests.

Change-Id: I7914a70686fa5cf680a1a431aa15faa66c40687d