Remove the local builds list from the executor client
This removes the local builds list (executor.builds) from the executor
client.

To make this work, we have to adapt a few tests which are still using
this list. To provide a similar functionality for those tests, the
ZuulTestCase class now provides a method getCurrentBuilds() that looks
up the builds from the active pipelines/queues.

Change-Id: If74bc7164d1156cdb7ba06f12868c60bf1617ea8