Update patch set 5
Patch Set 5:

Measurements in our system show that the getBuilds call takes ~400-500ms which adds that as a penalty to the event processing for every started build. At peak times we have ~200 build starts per minute which makes >3 build startups per second and with the event loop already being a bottleneck. I don't think we can take this into production in our system with this penalty without multiple schedulers.

Patch-set: 5
CC: Gerrit User 16068 <16068@4a232e18-c5a9-48ee-94c0-e04e7cca6543>