WIP: Add an API for ZK-backed objects
In order to put the pipeline state in ZooKeeper, we need objects
where the source of truth is in ZooKeeper, and as the pipeline
manager walks the queues and updates values, it writes that data
to ZK as it goes.  This will eventually allow us to reference
objects in ZK (so a build request could reference a build object
in the pipeline data structure in ZK).

To do that, we need to handle CRUD operations.  This base (or
mixin) class will allow us to perform these operations in the
context of a tenant write lock.  It handles retries and aborts
for object creation, updating, and loading.

Change-Id: I67b90f25197a57f28ca466562c2eaeaecd1494e0