Let zuul-web look up the live log streaming address from ZooKeeper
This removes the RPC call in zuul-web to look up the live log streaming
address from the build objects in the scheduler and instead uses the
build requests stored in ZooKeeper.

To make this work, the executor now stores its worker information
(hostname, log_port) on the build request when it locks the request.

This should reduce the load on the scheduler as it doesn't need to
handle those RPC calls anymore.

Change-Id: I359b70f2d5700b0435544db3ce81d64cb8b73464