Let zuul-web look up the live log streaming address from ZooKeeper
This removes the RPC call in zuul-web to look up the live log streaming
address from the build objects in the scheduler and instead uses the
build requests stored in ZooKeeper.

To make this work, the executor now stores its worker information
(hostname, log_port) on the build request when it locks the request.

Since the fingergw is using the same shared implementation to look up
the streaming address for a build's UUID, the implementation was moved
from zuul.rpclistener to zuul.lib.streamer_utils.

Additionally, the fingergw now has access to the component registry and
the executor api in ZooKeeper as both are needed to look up the
streaming address. To not create unnecessary watches for build requests
in each fingergw component, the executor api (resp. the
job_request_queue base class) now provides a "use_cache" flag. If
use_cache is set to False, no watches will be created. The cache is
enabled by default as this is still our main use case.

Overall this should reduce the load on the scheduler as it doesn't need
to handle the related RPC calls anymore.

Change-Id: I359b70f2d5700b0435544db3ce81d64cb8b73464