Delete build request by its path
This changes the executor_api.remove() method to accept a path rather
than a whole object. Doing so avoids that the executoro client first has
to look up the build request just to delete it in the next step.

As the merger and executor API share the same implementation, we have to
adapt the merger_api.remove calls as well.

Change-Id: I5d07a84ac1268891b9e7d2d2fd425a6f3bc0070b