Temporarily enqueue cycle changes for reporting
If we encounter a dependency cycle which is not permitted, the code
path is somewhat unusual: we create a dummy QueueItem and then
report it.  But we never actually enqueue that item.  This is a bit
of an anomaly as we work to put the pipeline state in ZK, so to make
things simpler, let's actually enqueue the item just long enough to
report it, then dequeue it.

That will have the effect of writing objects in ZK and deleting them,
but this shouldn't happen often.  It's better to optimize for
maintainability by having the process be the same rather than speed
or efficiency and needing to support a type of QueueItem that isn't
actually enqueued.

Change-Id: Ia1860ef7e59b3bc5d0eb635bb3fa253d2be53b4a