Periodicall run malloc_trim via apscheduler
Memory usage will increase over the lifetime of a scheduler instance.
This has largely to do with how the Python memory allocator works
internally. The reason for the increase is, that allocated arenas may
not be returned to the OS e.g due to memory fragmentation.

In order to reduce peak memory usage and make the actual memory usage
visible for monitoring, we will periodically (every 30min) execute
malloc_trim. With that we force libc to release free memory to the OS.

Change-Id: I8e5458f5344a5e76f9e9f6c895f832dd2f8e7d35