WIP: Don't clear dependent change cache
The onChangeUpdated method is called every time a change is updated.
Its purpose is to iterate over every change in the cache and, if the
updated change is a commit dependency (Depends-On) of the other change,
to set a flag telling the queue processor that changes' dependencies
are out of date and need to be refreshed.  The idea is that if this
happens, the queue processor should notice the dependencies no longer
match and dequeue it.

However, we also perform a similar check any time a change event arrives
at the scheduler.  We check if it's an update of a change in the queue,
and if so, we forward the event to the queue, and remove old versions
of changes.  Therefore by the time we get to processing a change which
depends-on a change ahead, the change ahead should already be gone anyway.
This means we can avoid the onChangeUpdated call and instead rely on
this behavior.  We still need to actually set the commit_needs in the
first place, so we do that when we enqueue the change.

At least, that was the idea.  This has two problems:

* We only remove a change on a patchset created event.  So if we have a
  3 change (C->B->A) series all using depends-on with the github driver,
  then if change B updates its PR message to point to Z instead of A,
  that won't trigger a dequeue of B.

  Maybe we can broaden that check to include dependencies being updated
  as well as patchset created.

* There's a pipeline option that allows changes to potentially remain
  in the pipeline even if they're updated.

TODO:
We might be able to solve both those problems by performing a lesser
version of what onChangeUpdated did, but have the queue processor do it
and only with the search set of changes in the queue.

Change-Id: Ifc40ff5583a14c7f015356b963e1ba354974e1c8