Web UI: Show pipeline types as icons
Add specific icons in front of pipelines names on the tenant
status page to help differentiate their characteristics. A mouseover
text provides a brief summary for each type of pipeline. Clicking on
the icon takes the user to the specific part of the documentation for
the type of pipeline.

Type <-> Icon mapping:

* timer trigger: OutlinedCalendarAltIcon (periodic jobs)
* independent: FlaskIcon (it is for experiments)
* dependent: CodeBranchIcon (integration pipeline)
* serial: SortAmountDownIcon (deploy all and in order)
* supercedent: BundleIcon (deploy latest)
* default: StreamIcon (icon used for pipelines in the build/buildset page)

Change-Id: Ia16993af70d133f15aad9a212737805c3623b76e