Add a test for github dependency updates
When C depends-on B depends-on A, and we update B to point at A1
instead of A, we dequeue the series since the dependency chain
was altered.  In gerrit, this may happen because altering the
depends-on footer necessarily requires a new patchset of B.  But
with github, the PR body can be altered without a new PS.  In that
case, the dequeue happens because of an explicit check of
dependencies.

Change Ifc40ff5583a14c7f015356b963e1ba354974e1c8 is a proposed fix
for an unrelated issue which probably would have broken this
functionality.  Make sure we test it so we don't regress.

Change-Id: I54fcd58e2be4d452dc5a986bb8b00a2ee067e53f