Add a test for dequeue-on-new-patchest
This setting (which is likely rarely used) instructs Zuul not to
dequeue an item merely because it has had a new patchset uploaded.

However, if another item depends on such a change, that other item
should still be dequeued because its dependencies are out of date.
Add a test for that so that future changes to the change cache don't
inadvertently alter this behavior.

Change-Id: Ib2e8504c37644e06d0ac0608fa7c56d092e2673d