Only refresh deps if change messages have changed
We only need to call the refreshDeps method if the Depends-On
list has changed.  That can only happen with a new patchset
(gerrit) or the PR body has changed (github et al).  Add a method
to determine if the PR body has changed so we can reduce the
times where we need to call this method.

Change-Id: Iaa50a274c29347397bc4e10e2c3cefc25e442879