[WIP] GUI: Add tenant selector to top menu
On a non whitelabeled setup, allow a user to jump from one tenant to
another without having to go back to the tenants page.
On a whitelabeled setup, make the tenant item non-clickable (the click
doesn't do anything anyway).

TODO: use a Dropdown rather than a Select, and replace the timezone
selector as well.

Change-Id: I94d27445c65ed5c3f8d02fae9d47d426528d2332