Fix race in test_data_return_child_from_retried_paused_job
It appears that Zuul may restart the lost build a little faster
than it used to during this test, and where we originally expected
to have either the old builds waiting or none, we may now already
have the new builds waiting.  We release the second, but not the
first, which is what actually would be waiting at that point (since
they should beth be retried).  Make this an unconditional release
so we avoid this problem (the sequence should be stable because
of the job graph dependency).

Change-Id: I5d6de4f53f4b918a1a276485519cde7e72ec1c29