Expose nodepool slot attribute
Nodepool now exposes a slot attribute which is set by the static
and metastatic drivers to provide a stable id for which "slot"
is occupied by a node on a host with max-parallel-jobs > 1.

Expose this as a variable to Ansible so that jobs can use it
to provide stable but non-conflicting workspace paths.

Depends-On: https://review.opendev.org/c/zuul/nodepool/+/811016
Change-Id: I07cea423df7811c1de7763ff48b8308768246810