Update patch set 12
Patch Set 12: Code-Review-1

We need a way to cleanup dequeued items from Zookeeper. The problem seems to be that we can't directly delete the item from Zookeeper on dequeue as e.g. the reported result is set after an item is dequeued. This will fail if it was already deleted on dequeue. The other option would be to explicitly delete an item if it is no longer needed.

Patch-set: 12
Reviewer: Gerrit User 27582 <27582@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1