Save and restore bundle with item in Zookeeper
Make sure that we restore a bundle so that all items in a dependency
cylce share the same bundle instance.

We also need take care of partially reported bundles as there might be
items that are no longer in the queue but are still references from the
bundle. Those items needs to be refreshed separately.

Change-Id: I33ba45a91f80dc24762f918f29d0e12d8d4df831