Fix deserializing change refs from cache
The change to a structured change cache key omitted an update to the
deserialization method for change objects, so when they were
deserialized from ZK, the fields such as "git_needs_changes" which
refer to other changes were converted to tuples instead of left
as string objects.

This change corrects that.

The easiest way to test this realistically is to use two schedulers,
where one populates the change cache and the other uses it.  We can
run a change through a pipeline in the first scheduler to populate
the cache, then pause that scheduler while running the change through
a second scheduler which will deserialize the change object from the
cache.  This test works at this point in v4 development because
the schedulers do not rely on shared pipeline state (we run the jobs
to completion before switching schedulers).

This necessitated a slight enhancement to the waitUntilSettled method
to only grab the run lock on a single scheduler when specified (so we
can verify that scheduler #2 is settled without regard to #1).

Change-Id: Idbd72dad3ca4fdeb60778a9dd0d4b2aef7ebebdb