Don't add node resources to nonexistent tenant
The periodic stats emitter may try to add up node resource usage
for nodes which belong to tenants the scheduler doesn't know about
yet because it's still starting up.  This causes the following
error:

Traceback (most recent call last):
  File "/usr/local/lib/python3.8/site-packages/zuul/scheduler.py", line 314, in runStats
    self._runStats()
  File "/usr/local/lib/python3.8/site-packages/zuul/scheduler.py", line 459, in _runStats
    self.nodepool.emitStatsTotals(self.abide)
  File "/usr/local/lib/python3.8/site-packages/zuul/nodepool.py", line 526, in emitStatsTotals
    self.addResources(resources_by_tenant[tenant_name],
  File "/usr/local/lib/python3.8/site-packages/zuul/nodepool.py", line 65, in addResources
    target[key] += value
TypeError: 'int' object is not subscriptable

First, we shouldn't initialize that dictionary to a defaultdict(int)
because it's actually a dict of dicts.  The (int) was leftover from
a previous implementation.

Second, we should just ignore nodes which belong to tenants or projects
we don't know about yet.  If we're supposed to track them, then we will
do so later once configuration is complete.

Change-Id: I552943ef9135041704f9849ef241c68d6b758a8a